wip(poet): Modularize Poet (WIP – VERY BROKEN)
