+++ src/Exceptions/LifecycleException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Log1x\Poet\Exceptions;
+
+use Exception;
+
+class LifecycleException extends Exception
+{
+    //
+}
